backend-python/requirements.txt
# FastAPI 및 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 데이터베이스 관련
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# 인증 및 보안
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 검증 및 직렬화
pydantic==2.5.0
pydantic-settings==2.1.0

# 로깅 및 유틸리티
loguru==0.7.2
python-dotenv==1.0.0

# 개발 도구
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 